Take a Note Application
This app helps you to store notes or important task that you might forget easily. The various features of this application are-
User Friendly.
Can store the task along with the title that the user want to give a particular task.
Saves the notes along with the date and time of the saving.
The technology stack used are-
Kotlin, Jetpack Compose, View Model, Room , Hilt, Dagger.